requirements/dev.txt
# all requirements for production, test, and dev
-r prod.txt
-r test.txt

django-debug-toolbar
pre-commit
black
#sphinx
#pytest-ordering